Adapted test scripts for exciton-phonon-coupling to latest code changes
