+++ test_scripts/Exc_Pho_Coupling/qcmd_4.py
@@ -1,4 +1,4 @@
-from wave_train.hamilton.coupled import Coupled
+from wave_train.hamilton.exc_pho_coupling import Exc_Pho_Coupling
 from wave_train.dynamics.qcmd import QCMD
 from wave_train.io.logging import TeeLogger
 from os.path import basename, splitext
@@ -15,7 +15,7 @@ def coupled_qcmd(batch_mode):
         logger = TeeLogger(log_file=my_file + ".log")
 
     # Set up the coupled exciton-phonon Hamiltonian for a chain
-    hamilton = Coupled(
+    hamilton = Exc_Pho_Coupling(
         n_site=41,                       # number of sites
         periodic=False,                  # periodic boundary conditions
         homogen=True,                    # Homogeneous chain/ring
